Factored quick notes out of main
